app/requirements.txt
PySimpleGUI==4.60.4
PyPDF2==3.0.0
PyMuPDF==1.22.3
reportlab==3.6.12
Pillow==9.3.0
